Challenge System

+++ src/index.ts
@@ -9,8 +9,10 @@ import Db from './db';
 import RedisCache from './RedisCache';
 import Selector from './model/Selector';
 import authorize, { AuthorizeResult } from './authorize';
+import { CHALLENGE_COMPLETION_COLLECTION } from './model/constants';
 
 const UNAUTHORIZED_RESULT = { message: 'Unauthorized' };
+const NOT_FOUND_RESULT = { message: 'Not Found' };
 
 const app = fastify({
   logger: true
@@ -28,6 +30,7 @@ const authenticate = async (request: FastifyRequest): Promise<DecodedIdToken | n
   authenticateHeader(request.headers.authorization);
 
 const unauthorized = (reply: FastifyReply) => reply.code(Error.CODE_NOT_AUTHORIZED).send(UNAUTHORIZED_RESULT);
+const notFound = (reply: FastifyReply) => reply.code(Error.CODE_NOT_FOUND).send(NOT_FOUND_RESULT);
 
 const cache = new RedisCache({
   host: config.redis.host,
@@ -52,7 +55,12 @@ app.get('/:collection/:id', async (request, reply) => {
   const selector: Selector = { collection, id };
 
   const authRes = await authorize(selector, token.sub, db);
-  if (authRes.type === AuthorizeResult.Type.NotAuthorized) return unauthorized(reply);
+  if (authRes.type === AuthorizeResult.Type.NotAuthorized) {
+    if (authRes.exists) return unauthorized(reply);
+    else return notFound(reply);
+  }
+
+  if (!authRes.read) return unauthorized(reply);
 
   reply.code(200).send(authRes.value);
 });
@@ -76,6 +84,7 @@ app.get('/:collection', async (request, reply) => {
 
 app.post('/:collection/:id', async (request, reply) => {
   const token = await authenticate(request);
+  console.log('token', token);
 
   if (!token) return unauthorized(reply);
 
@@ -85,18 +94,29 @@ app.post('/:collection/:id', async (request, reply) => {
 
   const value = request.body as object;
 
-  const authRes = await authorize(selector, token.sub, db);
-  if (authRes.type === AuthorizeResult.Type.NotAuthorized) {
-    if (authRes.exists) return unauthorized(reply);
-    if (!('author' in value)) return unauthorized(reply);
-    const author = value['author'];
-    if (!('type' in author)) return unauthorized(reply);
-    if (author.type !== Author.Type.User) return unauthorized(reply);
-    if (!('id' in author)) return unauthorized(reply);
-    if (author.id !== token.sub) return unauthorized(reply);
+  if (collection !== CHALLENGE_COMPLETION_COLLECTION) {
+    const authRes = await authorize(selector, token.sub, db);
+    if (authRes.type === AuthorizeResult.Type.NotAuthorized) {
+      if (authRes.exists) {
+        console.error('Exists but not authorized');
+        return unauthorized(reply);
+      }
+
+      if (!('author' in value)) return unauthorized(reply);
+      const author = value['author'];
+      if (typeof author !== 'object' || !('type' in author)) return unauthorized(reply);
+      if (author.type !== Author.Type.User) return unauthorized(reply);
+      if (!('id' in author)) return unauthorized(reply);
+      if (author.id !== token.sub) return unauthorized(reply);
+    } else if (!authRes.write) {
+      console.error('User tried to write to a collection they do not have write access to');
+      return unauthorized(reply);
+    }
   }
-  
-  const res = await db.set({ selector: { collection, id }, value });
+
+  const res = await db.set({ selector: { collection, id }, value, userId: token.sub });
+
+  console.log(res);
 
   if (res.type === 'error') {
     reply.code(res.code).send({ message: res.message });
@@ -116,8 +136,9 @@ app.delete('/:collection/:id', async (request, reply) => {
 
   const authRes = await authorize(selector, token.sub, db);
   if (authRes.type === AuthorizeResult.Type.NotAuthorized) return unauthorized(reply);
+  if (!authRes.write) return unauthorized(reply);
 
-  const res = await db.delete({ selector });
+  const res = await db.delete({ selector, userId: token.sub });
 
   if (res.type === 'error') {
     reply
